backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
opencv-python==4.10.0.84
pytesseract==0.3.13
Pillow==10.4.0
numpy==2.4.2
motor==3.6.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pdf2image==1.17.0
google-generativeai==0.8.4
authlib==1.4.0
